% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{set_dk_defaults}
\alias{set_dk_defaults}
\title{Set default fonts and set theme_dk() to be used in all plots}
\usage{
set_dk_defaults(base_font = "Inter Tight")
}
\arguments{
\item{base_font}{Inter Tight by default}
}
\description{
Set default fonts and set theme_dk() to be used in all plots
}
