% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_dk}
\alias{theme_dk}
\title{David Keyes's personal theme}
\usage{
theme_dk(base_family = "Inter Tight", base_size = 14)
}
\arguments{
\item{base_family}{Inter Tight by default}

\item{base_size}{14 by default}
}
\description{
David Keyes's personal theme
}
